refactor: 게임 결과 계산을 딜러에서 진행하도록 변경
